Bugfix: allow blockquote lines with just "> "
The nonSpecial parser conflated a skipPrefix parser that returned ""
with a parser that did not consume anything. It would attempt to parse
another character in the line if no output was returned, since
nonSpecial must consume at least one character, leading to a parse error
when the entire line was consumed by skipPrefix.

The fix is to be explicit about whether a prefix is expected, and to
stop using its output altogether.